DMM {MDS2021}	R Documentation

Fit Dirichlet Multinomial Models to count Data and Evaluate the Result with Adjusted ARI

Description

This function fits Dirichlet Multinomial Models (DMM) to a count matrix and then calculate agreement indices between any two partitions for the data set.

Usage


DMM(otu.count, k, group, randMethod = "MA",
    seed = 2021, trim = FALSE, threshold)

Arguments

otu.count
OTUs with count numbers or a data matrix. The samples are the rows.

k
an integer or a numeric vector. Specify the numbers of Dirichlet components to fit.

group
a numeric vector of class labels. The length of group should be the same as the number of samples.

randMethod
specifies the preferred external index to measure the agreement between the clustering result and the given group. Available indices are: "Rand", "HA" (Hubert and Arabie's adjusted Rand index), "MA" (Morey and Agresti's adjusted Rand index), "FM" (Fowlkes and Mallows's index), "Jaccard" (Jaccard index). By default, "MA" will be output.

seed
set.seed with the given number. By default, set.seed(2021).

trim
logical. By default, it's FALSE. If TRUE, please indicate the threshold.

threshold
trim the matrix with otu.count, which(colSums(otu.count) > threshold).

Value

A list of 3 containing k (the numbers of Dirichlet components), ARI (the corresponding adjusted Rand Indices) and cluster (the clustering results).

Examples


data(Martinez)
tree = Martinez$tree
otu.table = Martinez$otutable
otu.count = t(otu.table)
DMM(otu.count, group = as.numeric(as.factor(Martinez$sampleinfo)),
k = c(2,3,5), trim = T, threshold = 5)

[Package MDS2021 version 0.1.0 Index]