Smits {MDS2021}	R Documentation

Smits Data

Description

Smits data is a data set from the R package "MicrobiomeCluster" including OTU table, tree and sample seasonal information. The samples are from the early wet season and the late dry season.

Usage


data(Smits)

Format

A large list with 3 components which are tree, sampleinfo and otutable.

References

Smits, S. A., Leach, J., Sonnenburg, E. D., Gonzalez, C. G., Lichtman, J. S., Reid, G., Knight, R., Manjurano, A., Changalucha, J., Elias, J. E., Dominguez-Bello, M. G., and Sonnenburg, J. L. (2017). Seasonal cycling in the gut microbiome of the hadza hunter- gatherers of Tanzania. Science, 357(6353):802-806.

Examples


data(Smits)
table(Smits$sampleinfo)


[Package MDS2021 version 0.1.0 Index]